Removes dependency on /etc/resolv.conf
MS recently updated their docs about finding the host IP address.
Two are the documented ways:
1. 127.0.0.1 if net mode is mirrored or
2. default gateway if NAT.

As of WSL 2.2.1, /etc/resolv.conf contains an IP of type 10.0.0.0/8
if net mode is NAT but DNS tunneling is ON (which is the new default).
That breaks our host IP detection mechanism, because we relied on
finding an address of type 127.0.0.0/8 in /etc/resolv.conf nameserver
entry to decide whether we should try to talk to that address or the
default gateway.