[agent] Addresses duplication warning in tasks files
The structure of those files was already very close, and became even
more with the recent changes.
To avoid complain about the code duplication (which in this case is
intended and by design) I'm changing the order of the functions to
allow minimizing the amount of code covered by a `//nolint:dupl`
comment.