[agent] Clean-up helper setAgentUIDAndUpdateClientConf
We were attempting to be too flexible with allowing an empty conf.
With an empty conf we woudl'nt talk to Landscape, thus wouldn't have
an hostagentUID.
So, we can safely report "no need to notify" by returning an empty conf
result.
That simplifies the body, and we can get rid of the closure.
Also added a rollback in case a c.dump() fails.