[GUI] Updates SubscriptionStatusPage.create
The page itself doesn't change.
But we need to pass a boolean flag from the app to the model.
Here we use the Wizard route userdata param.
The existing tests don't quite change.
More tests are added, to assert that the
"Configure Landscape" button is not shown when globally disabled.

We intentionally leave the env var check to the top level. Tests would be harder that way
due global state.