Create the tarball in the download function
Well, it might now be doing more than it should,
But WSL imposes the precondition that nobody else holds the rootfs file open when registering the distro instance.
So we must ensure the file is closed by the time we call import.
This is the most readable way to do so IMHO.

We could have wrapped the file creation and the call to download in a closure. It would be more compliant to SRP,
but less readable.