Redesigns the Landscape view model
It was a single data holder with simple enough
interleaved logic, as there were not so many interactions between the fields.
The addition of the SSL Public key, only relevant to the self-hosted case,
although being just one extra field, would cause increase
in the logic interleave.
As the team agreed that the UX is better when the logic interleave mentioned above is clearly presented as subforms, the view model becomes simpler if it also
presents the data as separate elements,
(herein called data models individual classes inheriting from LandscapeConfig abstract/sealed class).

Input validation becomes a class-specific behavior, instead of an inter-field relationship. That's easier to maintain in the long run.

To reinforce that separation, tests are done in separate files, i.e.
the data models are tested in landscape_data_model_test.dart using parametric testing, which is great for data models
while the aggregator view model class (LandscapeModel)
is tested in its own file with more behavior-driven testing, as it's less about modelling data and more about managing behavior.