LandscapePage rewrite
We no longer use the ColumnLandingPage presented in page_widgets.dart.
Instead the entire layout is defined in the landscape_page.dart itself.
It's a complete rewrite.
Previously there was a clear column layout, now the layout is more horizontal.

Tab navigation required special attention: by default it would traverse the radios first, then moving to the subforms, focusing even on the invisible subforms.
To reinforce the relationship between a radio at left and the matching subform at right, I manually control the focus order to ensure:
- after tab focuses on a radio, the next tab-hit goes to its subform, if visible
- otherwise, goes to the next radio.

Parts of the UI that are expected to change less often as the subforms are filled in are wrapped in Selector widgets,
so we can model the entire LandscapeModel class as a ChangeNotifier without causing it to rebuild the entire UI when not all parts are affected by the change motivating a notification. Not a big performance concern, as the UI is not that big, but it's good to pay attention to details ;)

Finally, if the agent responds with a gRPC error, the error message is extracted and presented to the user as a SnackBar.