Prevents keys without values in the composed config strings.
This has been only problematic for ssl_public_key.
When the agent deploys that key without a value, wsl-pro-service invokes wslpath -au, which prints / :)

Maybe we should make wsl-pro-service smarter about that,
but its easier to get this mitigated here, as we know which fields we consider optional in the UI, we should prevent them to turn into keys without values.