One test case for UserProfileDir
We could in fact invoke wslpath with wrong arguments in different ways
there.
The test case only illustrates the case of the path being empty.
Note that the test currently passes, but for the wrong reason:
We try to stat a path under the CWD that doesn't exist, then we bail out.
Instead we should fail early, and that's what the upcoming fix will do.