Fix bug in test setup dir creation
File mode 0600 was not intended for this specific case,
as the setup step expects to be able to list and write into the
directory.
It was not noticed because the directory was previously created by
cloudinit.New.
Not that cloudinit won't create the directory if there is nothing to
write into, the bug was revealed.