writeAgentData now may wipe agent.yaml
If marshalConfig returns `nil, nil` there is no error but also nothing
for cloud-init to do. Thus, there should be no agent.yaml in disk.

So, we check for that condition and attempt to remove agent.yaml
from the cloudinit dataDir, if it exists. Missing file is not an error.