proservices.TestNew now needs to wait for cloud-init
We previously didn't bother with the contents of agent.yaml
But now that it will only exist after the registry notifications reach
the cloudinit component, we need to wait for that file to appear before
proceeding with any assertion.
And now we no longer expect to find an empty file, but a file that
matches what was set via the registry.