Test the store wrapper error translation
There is some non-negligible error translation in there
mostly due the somewhat involving error handling imposed by
syscall/dll_windows.
This adds a modest set of test cases to cover that logic.
For that to be approachable I extracted the error reporting
out of the proc call method, this way we don't depend on calling a DLL
function to test this part.
Quite artificial setup, but I always enjoy separating "computing" from
"doing".