Defines an API for network device monitoring
The Windows implementation is a thin wrapper around the Registry.
When discussing WS035 we agreed not to expose the registry, as
we might be required to find another, more specific, approach.
Thus the API methods don't resemble the registry, but a device
manangement API.
It's a building block to build the actual monitoring logic on top of it.

I'm placing it as a new package inside daemon because daemontestutils will need
the declaration of the DevicesAPI interface.