Remove the error channel from the daemon struct
That prevents an odd contract between daemon.serve and daemon.tryServeOnce
in which the first could exit early, thus the caller need to check for its return error,
and if nil, the serving goroutine would send its exit error via d.err, requiring the daemon to hold the channel.

With this change, serve returns the error channel which is the only source of truth, i.e.
the caller must unconditionally receive from that channel to know whether setting up the serving goroutine succeeded or not or if it's still running or not.

Finally, we use the ErrRestartDaemon sentinel error value to let the upper Serve restart the gRPC server.