Add a how-to deploy on multiple Windows machines
Landscape docs already feature a quite complete guide on how
to create WSL instances on Windows machines via Landscape API.

Here we take a subset of that documentation and put it in shape for
use with PowerShell (alongside Bash).

This guide's focus is to deploy WSL instances from a custom rootfs
with some cloud-config data for further customisation, using Landscape API
to do it at scale, with both Bash and PowerShell alternatives.

We keep a cross-reference to the existing Landscape docs, which are more
complete.