Makes the agent a single-instance program
By creating a lock file in the rootcmd.
This way we won't prevent other commands such as version or help to run.
Still early enough to prevent touching the disk or the network.

We try the lock file before setting up the logger so we don't mess with another process' logging.