Agent<->Landscape over TLS by default
Previously, if the Landscape server's SSL public key wasn't provided
we'd default to insecure credentials for gRPC.
That's not nice when the system trusts the server cert.
A browser would happily render the page as safe,
but the agent would fail to connect to the server's hostagent messenger.
We should assume TLS by default. If the server cert path is not provided,
load the system's cert pool and try to connect based on the system's chain of trust.
To avoid complicating the existing tests, we preserve the use of insecure credentials
via a context value. If set, we don't even check the cert path.
Otherwise, TLS by default.