Add a test case when server has a cert not trusted by the client's system
First we need to break `requireCertificate` into client and server bits, such that we can force the client to use TLS
while not having a defined ssl_public_key entry in the landscape.conf testdata file.
Client side is about using insecure credentials or not.
Server side is about generating new certificates or not.
Testdata landscape.conf dictates whether the client will read a self-signed cert or load the system's pool.

With that change, we add a test case in which
Client's landscape.conf doesn't have a ssl_public_key entry
but the test case still requires TLS, so system's cert pool is loaded.
As the server cert is a self-signed one, we expect this error message when gRPC attempts to connect over TLS:

 could not connect to Landscape server: rpc error: code = Unavailable desc = connection error: desc = "transport: authentication handshake failed: tls: failed to verify certificate: x509: certificate signed by unknown authority"