landscape-config --silent --register-if-needed
landscape-client (24.08+git6375-0ubuntu0) added the new
--register-if-needed flag for landscape-config, allowing it to refresh
an existing registration instead of unconditionally creating a new one
for a computer already registered with Landscape.

There are situations when wsl-pro-service subprocesses lansdcape-config
on an already registered computer. That creates a new registration and,
even though Landscape server has some logic to detect that the new
registration belongs to an existing computer, it loses data that takes
long to recover, such as package version information.

Since the landscape-client version in the latest-stable PPA is already
over that version we should be able to add this flag to wsl-pro-service
and prevent that behaviour.