Reports specific error if command is not found
We need to differentiate that error condition to determine
if we're failing because this is a "modern distro" instance.

Since the error type becomes part of the API a new test case
and the required mock changes were added as well.

Powershell only returns error code 1
We need to inspect the output text looking for something like

```
    + FullyQualifiedErrorId : CommandNotFoundException
```