No longer assume MS Store on install
Thus renaming into installFromWSLOnlineDistros()

gowsl.Install() effectively does `wsl --install`,
which could bring stuff from MS Store
as well as anywhere else, specially in the case of modern distros.

A helper function is added to handle modern distros.
We assume the old method is still the most common.
Modern setup is a fallback in case the launcher is not found.