improve titles and metas
search results for docs are primarily determined by the top-level header
and the meta descriptions used.

in the absence of meta descriptions, the first line of text after the
title is used in search results.

these changes attempt to add more descriptive titles which will be more
broadly understood when read in a search result.

in addition, meta descriptions are added so that we have more control
over how pages are described in search results.