add dev docs notice
it is possible that a regular user could visit a page intended for
developers.

for example, a user who wants to install UP4W to use it on their machine
ends up in the installation page for developers.

this adds a notice that these are not for general use.

in future, we might want to move them, but this is OK for now.