Prioritize writing metadata
This is unlikely to fail, but if that happened under a situation where
this data would be used to initialize an instance per Landscape request,
it would be an annoying experience, as the instance would be created and
initialised properly, except that the server would never tie such
instance to the activity that requested its creation. So making the
priority to write metadata first allows us to bail out early and retry
or whatever other recovery method may fit.