Update docs for new distro format
Change the command for running a distro:

For old format, <Distro>.exe is used.
We should default to wsl -d <Distro> for the new format.

Simplify the install command:

A distro can be installed with wsl --install <Distro> and
the -d flag is no longer required.

Add sections on running instances and starting directories:

With the old format, running ubuntu.exe started a user in Ubuntu home.
With the new format, wsl -d Ubuntu starts a user in the Windows CWD.

As this behaviour might not be expected, and deviates from non-terminal
methods to start an instance, it is worthwhile expanding on this topic
and indicating some options to the user.

Add clarifying notes on the new distro format:

Where appropriate, some context needs to be provided for the different
formats, especially as we are transitioning between them.

Update the distro list output:

The order is currently somewhat counterintuitive, as it reflects the
structure defined in a JSON file in the WSL repo.

To avoid a disparity between what the docs show and what the user sees,
the list should be updated.

Remove easily discoverable info from the how-to:

How-to guides are focused on adept users who want to get stuff done.

The outputs of commands and loading elements are more appropriate for
tutorials, where they can help a new user learn and gain confience.

Add more introductory material to tutorial for new users:

If this is a user's first experience with Ubuntu on WSL, some things
might trip them up.

The additional text is more explicit about Ubuntu being installed with
WSL, highlights the change of prompt that they can expect when running
an instance, and tells them how to exit the instance.

General improvements and fixes:

Aimed at improving flow and consistency, with removal of some redundant
information.

Fix codeblocks:

- ensures prompts aren't copied
- ensures output logs don't have copy button

Add install section on new distro format:

Previously we had content on importing an arbitrary tar.

We now support official tar.wsl releases so should document these
instead.

The advantages are briefly listed here, although this could be expanded
into a dedicated page at a later date.

External links are also provided for more information.

Spelling:

Fix spellings.

Update wordlist.

Ensure cloud-init setup before distro install:

If a distro has undergone user setup then cloud-init setup will fail.

To reduce error, the documentation has been updated to:

* Not assume a distro is installed before the guide is followed
* Replaces launch distro step with install distro setup
* Add a note about installing a distro without launching (also works)

General fixups for cloud-init guide:

Mostly minor changes to wording.

Change order of install methods:

Makes newer methods that are frequently used in the docs more prominent.