Updates CI to consider the app installer file
In the build-msix action we use the same technique that updates the package version in the manifest to also update the app installer file versions.

In the publish-msix workflow we create a draft release on GH uploading the app installer. We keep it in draft as we need to obtain and upload a signed version of the msix before completing the GH release.