Store Landscape FQDN as a URI
That implies careful handling of the case advertised in the Landscape docs: 'host:port'. URL parsing libs would treat that as 'scheme:path' instead.

We check for that case, and patch it to make it parseable.

Otherwise, we inject scheme and port if not provided.

Otherwise we assume the user knows what they're doing.

To be extra cautious we use `uri.host:uri.port` instead of `uri.authority` to prevent bundled user info.