Keep the defaulting logic in a single place
When sanitizing the Landscape server FQDN we're tempted to enforce
the defaults for incomplete inputs.
The user might still be typing stuff, so we should not rush to fix
all the things.

Let's keep the enforcement of defaults to the last minute, when
the user requests us to transform the data into a suitable Landscape
config instead of repeating ourselves or rushing.