Makes sure the receiveCommands() errors are checked in keepConnected()
We still want to take the decision about whether to bail out or not in
the same place we're already doing it.
The fact that the Landscape server returns us a hostagent UID before
validating the config we submitted makes it harder to detect such a
terminal condition.
We need to collect the errors emitted by the goroutine calling
receiveCommands() and bring them to keepConnected().

I'm now doing so by storing an error channel in the connection object.
While we could read it directly in the keepConnected() goroutine, there
is one very good reason for not doing so: my next PR wants to read the
same channel in a different place (the UI service) so that we can
respond the RPC with an error if the configuration was invalid.
Go channels are not broadcasting.

We have this goroutine monitoring connectivity. It previously signaled
when the connection was terminated so the keepConnected() method could
decide whether to retry or bail out. We'll use this place as the
broadcasting point. We now return the error collected from the
receiveCommands() error channel and propagate it via the ephemeral
connectionDone channel for the keepConnected() method. This same place
will be where we'll propagate the error to the UI service.

This place is special because it will also allow us to check that the
connectivity is good so the UI service can complete the RPC without
errors. This will be clearer in my next PR.