Landscape service to report ConnectionStatus()
That method blocks the caller until a gRPC connecitivity transitions
happen from the Connecting state to any other.
At Shutdown we collect errors from the receiveCommands() goroutine,
as something must have happened there, Recv() would have reported
and we can distribute here into the ephemeral connectingDone channel
as well as the longstanding connErrs channel, that feeds the
ConnectionStatus() method.

It's only expected usage is on the Landscape notifier arranged by the
proservices.New method, so that when a new Landscape configuration is
submitted via the UI we have ways to return errors in case that
configuration is erroneous in some way only detectable by actually
trying to talk to the Landscape server.