ui.ApplyLandscapeConfig should wait for a connection result
To allow the GUI to handle errors when submitting a Landscape config,
the UI Service should delay replying the relevant RPC until we have a
confirmation that the provided configuration worked, meaning we're able
to keep a command stream open with the server.
Also, since the config module doesn't reapply a config that doesn't
change anything, we should treat that as an error, otherwise we'd block
the RPC forever waiting a notification that won't ever come.

The introduction of the channel and the Stop() method required that the
proservices manager hold a uiservice object by pointer, otherwise we
could get double-close errors.

TODO: config module to report an error when the submitted config is unchanged.