Landscape service to notify listeners ...
... of interesting connectivity events related to the receiving commands stream.

This is the trickiest part of this changeset. Finding the exact sequence of events that allow notifying the UI at the right times was quite complicated due the fact that a successful handshake() doesn't mean the configuration is valid.

The Landscape server immediately terminates a connection when the WSL feature flag is turned off. That's great! Easy to handle.

When the configuraiton is invalid, things are much more complicated. We receive an UID, which means the receiveCommands() loop will have a first succesful Recv followed by a successful command execution, and at a later point during a second Recv we find that the configuration data is invalid and have the connection terminated.
Finding a generic way to handle those terminations and still being capable of declaring a connection good (so the UI RPC can be successfully terminated) proved to be complex.

The Landscape team is commited to simplify things from the server side: only reply with a UID after the config was validated, terminating the connection immeditaly with InvalidArgument otherwise. See LDENG-3037.

But we may be required to connect to older versions of the server (as we are now, for example), so most of this changeset is to couple with this condition.

That's the reason why the gRPC connectivity monitoring goroutine is so full of comments now ;)

Despite the complexity of the code changes, I'm doing the commit now before creating new test cases for the Landscape service to prove that those changes don't change the existing behaviour significantly from the perspective of the existing tests: they pass once we pass `nil` as the connection listener in the calls to landscape.New().