LandscapePage display an error alert
The alert is not dismissible by default because we expect the user
to take some action. The action can be just go back to editing their
config, dismissing the alert, or even more dangerously,
forcing submitting the erroneous config as is.

Who knows if the user is not submitting a config for a to-be-deployed
server for testing for example.

This allows the user to see some error cases immediately without getting
completely on their way, preventing advanced use.