More comprehensive checks for when to disable Landscape
Besides rearranging the code for readability, we include DNS failures.
We certainly don't want to disable Landscape if we fail to resolve names due not
being able to talk to a DNS server in the first place.
So this goes a bit more invasively and checks for messages that
are generated when talking to a DNS server works but the server can't find such host.