Overlays a spinner on top of the 'Register' button
That required a small refactoring in the NavigationRow widget, to allow passing a Widget as the 'next' parameter, instead of assuming it willl always be Text.

Just replacing the Text with a CircularProgressIndicator would resize the button. That's somewhat disturbing.
The solution is to preserve the text in the layout but make it invisible and overlay the spinner on top.