Section about UP4W components deployment and updates
Explaining briefly the format of distribution for the UP4W components
and how they get updated.

Specifically mentions the appinstaller file for Windows 10, which
doesn't get associated with the MSIX package if the user installs the
MSIX directly. On Windows 11 all installation methods are granted
automatic updates tracked by the OS. On Windows 10 only if installed
from the MS Store or via the Appinstaller file.