Remove tools module from go.work
None of the other modules depend on the tools module. The purpose the
tools module is to define the versions we want to use for tools like
golangci-lint and govulncheck in the CI. Updating golangci-lint to v2
requires updating the Go version of the tools module to 1.24.0. If the
tools module is part of the workspace, all modules in the workspace will
be built with Go 1.24.0. Removing it allows us to keep building the
other modules with lower Go versions.