Prevents race in t.Cleanup() for tests relying on Landscape mock server
The existing construct is inherently racy: it's possible to have other
goroutines still writing to the slog while the t.Cleanup() function
attempts to read from its underlying buffer.

To prevent this condition I implemented a custom slog Handler that wraps
the slog.TextHandler but only proceeds with writing a record to the log
after acquiring a mutex, which is also used on reading in the
t.Cleanup().