Leverages ubuntu_logger.dart.
Logging interesting events in the UI to its own file.
Including unhandled errors propagated by FlutterError and
PlatformDispatcher.
It also adds an invasive route observer to the Wizard router to
log the route navigation.
This app is really small, but has some requirements about navigation.
It's good to have that recorded in case users report bugs so we can
detect what should be impossible navigation.