Unconditionally add the App Installer file as release asset
I realized Windows 10 users need that file easily discoverable as that
must be the preferred way to install the MSIX with guaranteed
auto-updates if not by the MS Store.

Thus we need to make it visible whether it's a pre-release or stable
release.

Importantly, that doesn't change the `AppInstaller.Uri`, which remains
pointing to the latest stable release or to the latest version of that
file stored in the wiki.