Adds a distro database test related to reporting unamanged instances
The test sets up a few Ubuntu and one non-Ubuntu instances
(differentiated by their os-release file).

Some are added to the database (mimic'ing the effect of wsl-pro-service
proactively reporting the instance to the agent) while others are left
out of the DB (mimic'ing instances that don't have that service
enabled).

We then assert that no non-Ubuntu and no managed instances
are present in the unamanged list, in acordance with the spec WS043.