Pins Windows SDK version where relevant
We previously pinned 10.0.22621.0 because the version 26200
had issues with debug builds in CI (missing symbols).

CI is now pushing 10.0.26100.0 by default (that's the only SDK version
installed).

If that doesn't come with any noticeable regression, then the upgrade is
worth it. Otherwise we'd need to force CI to provide the 10.0.22621.0.