Skip clang-tidy on the Flutter GUI but still generates the compilation database
What we need in the end is just the compilation databases (compile_commands.json) for every project in the solution that relies on native C++ code, such as the storeapi.DLL, the ubuntu-pro-agent-launcher and the native side of the GUI.
TICS controls a run of its own clang-tidy version, ignoring our local configuration and CI run results.

So what I'm doing here is essentially tricking msbuild (ran by Flutter) into preparing for clang-tidy to run (creating the compile_commands,json) but run our local script instead of the linter.

That's needed because VS 2022 ships clang-tidy 19, which cannot compile code that depends on MSVC STL implementation of <compare>. There is a circular requires-clause in that header that clang fails to compile. The issue was fixed in clang 21, but that's not yet distributed with VIsual Studio so we cannot rely on it yet (unless we ship LLVM ourselves).

TICS also relies on LLVM 19, so let's see how far we can get.