Git tree not necessily a monorepo checkout
We previously assumed that if we are under a git worktree it means
we are in a monorepo checkout and we should recompute version.
But we could well be under a git-ubuntu checkout, for which we only
have the source package contents, not the entire monorepo.

To know whether we should or could recompute version, we then
check for git and for the existence of the compute_version.go helper.