Invokes TiCS twice
One for Linux and other for Windows

I just learned that the acronym UP4W seen in TiCS dashboard is a repo or
project group, not the project itself, which is split in two projects.
The correct way to report QA is then submitting it against the proper
project name as seen in the dashboard.
The proper fix is likely to split CI entirely so we process
wsl-pro-service differently from the Windows parts, but since they have
parts in common, my first attempt is to report everything all together
twice. Let's see how it goes.