v1-only: Let's not create the LandscapeConfig registry key by default.
Since v1 is meant to not advertise Landscape support, having us creating
a seemingly unrelated registry key automatically can be confusing for
the user. Let's not do it. Hopefully this will be the only code change
between v1 and the main branch by the time we freeze for the release.