fix(ci): QoL - no partial codecov reports
I remember relying solely on the Cobertura format for codecov was
yielding subpar results because it didn't handle partial coverage in
branches. Yet, that's not enough of a reason to preserve this step
issuing partial uploads to codecov when we could just include the gocov
binary format as part of the collected files from the test matrix.