Assume the MSIX bundle name won't contain the version
We want the URl
https://github.com/canonical/ubuntu-pro-for-wsl/releases/latest/download/UbuntuProForWSL.msixbundle
to always work and always point to the latest **stable** release of this
app.
If that file was named including the version, that wouldn't work.
Thus we assume the people uploading the bundle to the release won't add
the version to its name.
To prevent accidents, we assume the MSIX bundle file won't ever contain
the version in its filename

---

UDENG-8328