Do not terminate an instance after Landscape install
While that helped with the RR case, it pessimized everything else:
it's now possible to have wsl-pro-service doing something (trying to run
landscape-config for example) and we forcibly kill the instance before
it completes.

As I'm pushing a new upload to RR solving the wsl-pro-service part, this
shutdown is no longer necessary.