fix sitemap regression introduced in 1.3.0
The 1.3.0 starter pack introduced a regression that duplicated versions
in sitemaps, leading to malformed URLs. 1.3.0 introduced fetching the
canonical URL from a RTD environment variable. In versioned docs, this
URL includes the version, which means that the previous logic to set the
sitemap_url_scheme is no longer necessary and duplicates the version.

This change should hopefully fix the sitemaps.