CSS updates
* move .css files into dedicated CSS dir
* change default background colour of dropdowns

The default bg colour of dropdowns looks bad and is inaccessible.
We got around this before by specifying Primary for the dropdown colour,
but this creates somewhat of a clash with out default theming.
It also means that maintainers need to remember to specify "primary".

Following the Enterprise Store docs, I've added a small CSS rule to
improve the default syling of dropdowns.